+++ examples/pothos/src/index.ts
@@ -1,0 +1,5 @@
+import { createServer } from 'http'
+import { yoga } from './yoga'
+
+const server = createServer(yoga)
+server.listen(4000)