+++ packages/graphql-yoga/__integration-tests__/subscription.spec.ts
@@ -1,7 +1,7 @@
 import { createYoga, createSchema } from 'graphql-yoga'
 import { createServer } from 'http'
-import EventSource from 'eventsource'
-import getPort from 'get-port'
+import { fetch } from '@whatwg-node/fetch'
+import { AddressInfo } from 'net'
 
 describe('subscription', () => {
   test('Subscription is closed properly', async () => {
@@ -48,17 +48,27 @@ describe('subscription', () => {
     })
     const server = createServer(yoga)
     try {
-      const port = await getPort()
-      await new Promise<void>((resolve) => server.listen(port, resolve))
+      await new Promise<void>((resolve) => server.listen(0, resolve))
+      const port = (server.address() as AddressInfo).port
 
       // Start and Close a HTTP SSE subscription
-      await new Promise<void>((res) => {
-        const eventSource = new EventSource(
+      await new Promise<void>(async (res) => {
+        const response = await fetch(
           `http://localhost:${port}/graphql?query=subscription{foo}`,
+          {
+            headers: {
+              Accept: 'text/event-stream',
+            },
+          },
         )
-        eventSource.onmessage = (ev) => {
-          eventSource.close()
-          res()
+        expect(response.status).toBe(200)
+        expect(response.headers.get('content-type')).toBe('text/event-stream')
+
+        for await (const chunk of response.body!) {
+          const chunkStr = Buffer.from(chunk).toString('utf-8')
+          if (chunkStr) {
+            res()
+          }
         }
       })
       resolve()