+++ packages/event-target/typed-event-target/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @graphql-yoga/typed-event-target
 
+## 1.0.0-next.0
+
+### Major Changes
+
+- [#1761](https://github.com/dotansimha/graphql-yoga/pull/1761) [`b2407c6a`](https://github.com/dotansimha/graphql-yoga/commit/b2407c6addab136e3390bd4efa1fbbad7eb8dab8) Thanks [@ardatan](https://github.com/ardatan)! - **BREAKING**:
+
+  - Drop `TypedEvent` in favor of [`CustomEvent`](https://developer.mozilla.org/en-US/docs/Web/API/CustomEvent/CustomEvent)
+  - Use `@whatwg-node/events` as a ponyfill instead of `@whatwg-node/fetch`
+
 ## 0.1.1
 
 ### Patch Changes